requirements.txt
python-dotenv~=0.17.1
psycopg2~=2.8.6
gensim==4.0.1
pymystem3==0.2.0
nltk==3.6.2